requirements.txt
audioop-lts==0.2.2
audioread==3.1.0
backports.functools-lru-cache==2.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
decorator==5.2.1
filelock==3.20.3
fsspec==2026.1.0
idna==3.11
indic-transliteration==2.3.75
Jinja2==3.1.6
joblib==1.5.3
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
openai-whisper==20250625
packaging==25.0
platformdirs==4.5.1
pooch==1.8.2
pycparser==2.23
pydub==0.25.1
Pygments==2.19.2
regex==2026.1.15
requests==2.32.5
rich==14.2.0
roman==5.2
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.9.0
shellingham==1.5.4
soundfile==0.13.1
soxr==1.0.0
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
toml==0.10.2
torch==2.9.1
tqdm==4.67.1
typer==0.21.1
typing_extensions==4.15.0
urllib3==2.6.3
webrtcvad==2.0.10
